Fix SparseLengthsSumI8

+++ lib/Backends/Interpreter/InterpreterNodes.cpp
@@ -1932,7 +1932,7 @@ void InterpreterFunction::fwdSparseLengthsWeightedSumInst_I8Impl(
   out->zero();
 
   auto IH = indices->getHandle<int64_t>();
-  auto LH = lengths->getHandle<int64_t>();
+  auto LH = lengths->getHandle<int32_t>();
 
   size_t segments = lengths->dims()[0];
   size_t totalLength = 0;
@@ -1957,7 +1957,7 @@ void InterpreterFunction::fwdSparseLengthsWeightedSumInst_I8Impl(
   size_t curIdx = 0;
   for (size_t i = 0; i < segments; i++) {
     std::vector<float> accum(lineSize, 0.0f);
-    for (size_t j = 0; j < LH.raw(i); j++) {
+    for (int32_t j = 0; j < LH.raw(i); j++) {
       float weight = dequantize(WH.raw(curIdx), TQP(weights));
       size_t offsetIn = IH.raw(curIdx) * lineSize;
       for (size_t k = 0; k < lineSize; k++) {
